Refactor service worker code, start using Sentry releases
